Fixed: #66 Freezed Y-Rotation on Players Rigidbody
This is the absolute way to stop the Player from rotating after collisions. Another less brute method would be to increase the angular drag value of the Rigidbody Component. However, as long as we dont need the Player to be rotated by other objects we can leave this well enough alone.