Added: #32 Lockpick System /w SoundFX, Fixed: #76 Exanded door trigger, Fixed: #87 Shrinked down torches and chains
I'm sorry for not having split the Issues into single commits, but due to the time pressure of the Games Bavaria Vernissage tomorrow I simply forgot it. The biggest part of this commit is the new Lockpick System with SoundFX for fail and succes in the unlocking process. There should be HUD instructions for how to play, but for now I only use Debug messages. So please look at the Console for further information!