Updated: #121 Shadow Script
Smoothed out detection values. Now the script calculates the average color of an adjustable area (ideally the area covered by the player) and with an adjustable number of probes. There is still a problem with the edges of the lightmap tiles, which causes unwanted values.
Additionally added getter method to GameController.