Changed: #73 Texturing of protagonist ~ 80%
Texturing of protagonist is nearly done.
Please do not look at her face, because she seems like suffering down
syndrome.
Because we will never see her front face it is not that horrible.
I tried raising the texture resolution, but it is not possible, due to a
"low res" uv mapping-picture - therefore the textures look a little bit
washed out.
This is not a professional course in the university therefore I think it
is not that bad when it is not 100% perfect.